backend/requirements.txt
asgiref==3.11.1
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg2==2.9.11
sqlparse==0.5.5
tzdata==2025.3
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1
reportlab==4.1.0
